Re-architectured project parameter in its own class HTMLParameterProject
